Improve isolation of the render context
Unset the arguments array and redirect stdin to /dev/null so it can't be consumed by code in template expressions, which would result in loss of data.

Also fix and add more early returns in case of errors. Some errors were getting lost or could still slip through, making it difficult to debug issues in templates.