Fix truncation of trailing whitespace
Command Substitution truncates trailing whitespace, that's a general sh issue, so it's necessary to guard against it. This was a problem in the quoted expressions that need to be escaped before printing to stdout. The test for this case has also been fixed.

Now all tests are passing, and the program can be considered feature-complete.