Use async pipeline to avoid excessive buffering
Assigning the output of a command to a variable requires buffering the entire output, which can lead to reading large files into memory. Change to an asynchronous pipeline instead. Now the commands run concurrently and flush the lines of text as they come, instead of buffering the entire result.

There should be no negative impact on status code propagation thanks to the new Pipeline Error Propagation Protocol created specifically for this use-case.