Do not include literal hashes
The idea behind the hashes is to have unique tags that are almost impossible to collide with any kind of real data. Having them written literally in the source file (which is real data) in the order they are expected to appear when used does not help that case. Generate new hashes and add some non-significant quotations in them so they are never written literal anywhere.